refactor : 도메인 구분을 위한 participant 패키지 분리
