+++ src/main/java/blackjack/domain/participant/Dealer.java
@@ -1,21 +1,22 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
+import blackjack.domain.Card;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
-public class Player {
+public class Dealer {
 
+    private static final int DRAWABLE_MAX_SCORE = 16;
     private static final int BLACKJACK_SCORE = 21;
 
     private final List<Card> cards;
 
-    public Player(List<Card> cards) {
+    public Dealer(List<Card> cards) {
         this.cards = new ArrayList<>(cards);
     }
 
     public boolean isDrawable() {
-        return calculateScore() <= BLACKJACK_SCORE;
+        return calculateScore() <= DRAWABLE_MAX_SCORE;
     }
 
     public int calculateScore() {
@@ -33,8 +34,8 @@ public class Player {
         return score;
     }
 
-    public List<Card> getCards() {
-        return Collections.unmodifiableList(cards);
+    List<Card> getCards() {
+        return cards;
     }
 
     public void add(Card card) {
